+++ src/components/Header.js
@@ -1,48 +1,107 @@
-import React, { forwardRef, useEffect, useRef, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import { useHistory, useParams } from "react-router-dom";
 import { useHotkeys } from "react-hotkeys-hook";
 
 import { makeFriendly } from "./../utils/num.js";
 
+import {VscColorMode} from "react-icons/vsc";
+
 import "./header.css";
 
-export default function Header({ subCount }) {
+export default function Header() {
 	const { subreddit } = useParams();
 	const history = useHistory();
 	const [selectMenuOpen, setSelectMenuState] = useState(false);
 	const toggleSelectMenu = () => setSelectMenuState(!selectMenuOpen);
 	const closeSelectMenu = () => setSelectMenuState(false);
 
-	useHotkeys("s", toggleSelectMenu);
-
 	const sel_subreddit = (sub) => {
 		closeSelectMenu();
-		// change url.
 		// sub includes "r/"
 		history.push("/" + sub.slice(2));
-		// setSubreddit(sub.slice(2));
 	};
+
+	const [imgSrc, setImgSrc] = useState(null);
+	const [subCount, setSubCount] = useState();
+	const [active, setActiveCount] = useState();
+	const [desc, setDesc] = useState();
+	const img = useRef();
+
+	useHotkeys("s", toggleSelectMenu);
+
+	// get the info about the subreddit. do when sub changes.
+	useEffect(() => {
+		// reset img display because some subreddit's icons may not load.
+		img.current.setAttribute("style", "display: none;")
+		// sub_info: https://www.reddit.com/dev/api/#GET_r_{subreddit}_about
+		// data.public_description, header_img, allow_galleries, wiki_enabled, active_user_count, icon_img
+		// allow_videos, submission_type, created, spoilers_enabled, over18
+		fetch(`https://www.reddit.com/r/${subreddit}/about.json`)
+			.then((res) => res.json())
+			.then(({ data }) => {
+				// sometimes there is no icon_img
+				// shit 403 on data.community_icon
+				setImgSrc(data?.icon_img || data?.community_icon);
+				setSubCount(data?.subscribers);
+				setActiveCount(data?.active_user_count);
+				setDesc(data?.public_description)
+				// todo: dont show the icon if we cant load it.
+			});
+	}, [subreddit]);
+	// todo: public_description_html vs public_description
+
+	const toggleTheme = () => {
+		/*
+		--primary: #edf6f9;
+		--text: #1c1412;
+		--bg: #1d1f21;
+		--fg: #C5C8C6;
+		*/
+		const r = document.querySelector(':root');
+		const rs = getComputedStyle(r);
+		const [primary, text] = [rs.getPropertyValue('--primary'), rs.getPropertyValue('--text')]
+		if (primary === "#edf6f9" && text === "#1c1412") {		
+			r.style.setProperty('--primary', '#1d1f21');
+			r.style.setProperty('--text', '#C5C8C6');
+		} else {
+			r.style.setProperty('--primary', '#edf6f9');
+			r.style.setProperty('--text', '#1c1412');
+		}
+	}
+
+	useHotkeys("t", toggleTheme);
 
 	return (
 		<header>
 			{/*welcome to The Open Source reddit client focused on browsing{" "}*/}
 			<span>
-				<span onClick={toggleSelectMenu}>
+				<span onClick={toggleSelectMenu} title={desc}>
 					<img
-						src="https://styles.redditmedia.com/t5_2szyo/styles/communityIcon_x3ag97t82z251.png?width=256&s=33531dceba6466953aadef3073f36cfc2e267175"
-						alt="showerthoughts subreddit logo"
+						src={imgSrc}
+						alt="subreddit logo"
 						width="50px"
 						height="50px"
+						style={{display:"none"}}
+						ref={img}
+						onLoad={()=>img.current.setAttribute("style", "display: inline-block;")}
 					/>
 					<p className="banner">r/{subreddit}</p>
 				</span>
 				{![null, NaN, undefined].some((v) =>
 					Object.is(subCount, v)
 				) && (
+					<>
 					<p className="members" title={subCount}>
 						{makeFriendly(subCount)} members
 					</p>
+					<p className="members" title={active}>
+						{makeFriendly(active)} active
+					</p>
+					</>
 				)}
+				{/*<span className="desc">
+					{desc}
+				</span>*/}
 				{/* yup nice now i just need to know a good way to make forms in react! THIS SHOULD BE ITS OWN COMPONENT TODO: ADD FOCUS ON USEFFECT*/}
 				{selectMenuOpen && (
 					// NICE  better way to pass props
@@ -50,6 +109,9 @@ export default function Header({ subCount }) {
 						{...{ sel_subreddit, subreddit, closeSelectMenu }}
 					></SubredditSelect>
 				)}
+			</span>
+			<span className="theme-switch">
+				<VscColorMode onClick={toggleTheme}/>
 			</span>
 		</header>
 	);