+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## CHANGELOG
 
-1. fixed header height.
-2. fixed the react-stack-grid images overlapping issue.
-3. found bugs.
+1. fixed broken sub icon being show.
+2. added active sub count.
+3. added sub description.
+4. added dark switch button (press 't' to switch).